chore: renamed files in tasks/actions to avoid repeating action
