+++ src/features/tasks/actions/index.ts
@@ -1,1 +1,1 @@
-export * from './actions';
+export * from './core';
